Fallback to other PHP cli binaries if the selected one doesn't exist in the jail

+++ server/plugins-available/cron_jailkit_plugin.inc.php
@@ -382,9 +382,10 @@ class cron_jailkit_plugin {
 		$tpl->setVar('use_php_path', false);
 		$tpl->setVar('use_php_alias', false);
 
-		$php_bin_dir = dirname($this->parent_domain['php_cli_binary']);
 
 		if(($this->parent_domain['server_php_id'] > 0) && !empty($this->parent_domain['php_cli_binary'])) {
+			$php_bin_dir = dirname($this->parent_domain['php_cli_binary']);
+
 			if(preg_match('/^(\/usr\/(s)?bin|\/(s)?bin)/', $php_bin_dir)) {
 				$tpl->setVar('use_php_path', false);
 				$tpl->setVar('use_php_alias', true);
@@ -399,9 +400,19 @@ class cron_jailkit_plugin {
 				$app->log("The PHP cli binary " . $this->parent_domain['php_cli_binary'] . " is not available in the jail of the web " . $this->parent_domain['domain']  . " / cronjob_id: " . $this->data['new']['id']  . ". Check your Jailkit setup!", LOGLEVEL_DEBUG);
 				$tpl->setVar('use_php_path', false);
 				$tpl->setVar('use_php_alias', false);
-				if(is_link($this->parent_domain['document_root'] . '/etc/alternatives/php'))
-				{
-					unlink($this->parent_domain['document_root'] . '/etc/alternatives/php');
+
+				if(!empty($app->system->get_newest_php_bin($this->parent_domain['document_root'] . $php_bin_dir))) {
+					if(is_link($this->parent_domain['document_root'] . '/etc/alternatives/php'))
+					{
+						unlink($this->parent_domain['document_root'] . '/etc/alternatives/php');
+					}
+
+					$fallback_php = $app->system->get_newest_php_bin($this->parent_domain['document_root'] . $php_bin_dir);
+					$fallback_php_bin = str_replace($this->parent_domain['document_root'], '', $fallback_php);
+
+					symlink($fallback_php_bin, $this->parent_domain['document_root'] . '/etc/alternatives/php');
+
+					$app->log("Found " . $fallback_php_bin . " as a fallback in the jail of ". $this->parent_domain['domain'], LOGLEVEL_DEBUG);
 				}
 			} else {
 				if($app->system->get_os_type() == "debian" || $app->system->get_os_type() == "ubuntu") {
